Move Rarity.java to its own package
