+++ even-more-fish-plugin/src/main/java/com/oheers/fish/baits/Bait.java
@@ -7,7 +7,7 @@ import com.oheers.fish.config.BaitFile;
 import com.oheers.fish.config.messages.ConfigMessage;
 import com.oheers.fish.fishing.items.Fish;
 import com.oheers.fish.fishing.items.FishManager;
-import com.oheers.fish.fishing.items.Rarity;
+import com.oheers.fish.fishing.items.rarities.Rarity;
 import com.oheers.fish.utils.ItemFactory;
 import org.bukkit.Location;
 import org.bukkit.OfflinePlayer;
