+++ even-more-fish-plugin/src/main/java/com/oheers/fish/commands/AdminCommand.java
@@ -20,7 +20,7 @@ import com.oheers.fish.config.messages.ConfigMessage;
 import com.oheers.fish.config.messages.Messages;
 import com.oheers.fish.fishing.items.Fish;
 import com.oheers.fish.fishing.items.FishManager;
-import com.oheers.fish.fishing.items.Rarity;
+import com.oheers.fish.fishing.items.rarities.Rarity;
 import com.oheers.fish.permissions.AdminPerms;
 import de.tr7zw.changeme.nbtapi.NBT;
 import net.md_5.bungee.api.chat.*;
