+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/configs/CompetitionFile.java
@@ -8,7 +8,7 @@ import com.oheers.fish.competition.Bar;
 import com.oheers.fish.competition.CompetitionType;
 import com.oheers.fish.config.ConfigBase;
 import com.oheers.fish.fishing.items.FishManager;
-import com.oheers.fish.fishing.items.Rarity;
+import com.oheers.fish.fishing.items.rarities.Rarity;
 import dev.dejvokep.boostedyaml.block.implementation.Section;
 import org.bukkit.Bukkit;
 import org.bukkit.Sound;
