+++ even-more-fish-plugin/src/main/java/com/oheers/fish/fishing/items/Fish.java
@@ -8,6 +8,7 @@ import com.oheers.fish.api.reward.Reward;
 import com.oheers.fish.config.FishFile;
 import com.oheers.fish.config.messages.ConfigMessage;
 import com.oheers.fish.exceptions.InvalidFishException;
+import com.oheers.fish.fishing.items.rarities.Rarity;
 import com.oheers.fish.selling.WorthNBT;
 import com.oheers.fish.utils.ItemFactory;
 import de.tr7zw.changeme.nbtapi.NBT;
